float in return?

+++ python_variable_annotations/7-to_kv.py
@@ -5,4 +5,4 @@ from typing import Union, Tuple
 
 def to_kv(k: str, v: Union[int | float]) -> Tuple[str, float]:
     '''Returns a tuple'''
-    return (k, v * v)
+    return (k, float(v * v))
